Fix #1: "atomically" read /proc files
This commit addresses the problem of reading files in `/proc` with a
buffered reader, e.g. `std::ifstream`.  With a buffered reader, a file
may be read in by multiple partial reads.  Between these partial reads,
the process may spend time on other work, e.g. parsing.  Within that
time frame, the file itself -- which is just a memory mapping of
kernel information -- may change.  A partial read *after* such a change
may cause spurious, non-repeatable readings and is basically a race
condition.

To circumvent this obstacle, this commit attempts to read the files in
`/proc` *atomically*, that is it performs a single read of the entire
file into a local buffer.  After reading of the entire file, that buffer
can be accessed and parsed (partially or entirely) without causing
races.

There is a notable issue with this attempt, though.  The kernel system
calls to read from a file, i.e. `read(2)` and `pread(2)`, are allowed to
read *less* bytes than requested.  This is *not* an error but it signals
to the calling process that successive reads are required.  The code
acknowledges this behavior and performs multiple successive reads of the
file, should it be necessary.  This means, in fact, that this is still
racy!  However, immediately querying reads for the entire file should
significantly reduce the time frame in which the file may change and
reduce the frequency of races.